Modernize database setup in demo app
