+++ Documentation/RxGRDBDemo/RxGRDBDemo/AppDelegate.swift
@@ -10,7 +10,17 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     let disposeBag = DisposeBag()
     
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        dbPool = try! openDatabase(application)
+        try! setupDatabase(application)
         return true
+    }
+    
+    private func setupDatabase(_ application: UIApplication) throws {
+        let documentsPath = NSSearchPathForDirectoriesInDomains(.documentDirectory, .userDomainMask, true).first! as NSString
+        let databasePath = documentsPath.appendingPathComponent("db.sqlite")
+        dbPool = try AppDatabase.openDatabase(atPath: databasePath)
+        
+        // Be a nice iOS citizen, and don't consume too much memory
+        // See https://github.com/groue/GRDB.swift/#memory-management
+        dbPool.setupMemoryManagement(in: application)
     }
 }
